ENH: moved TargetCreationWidget to widgets and CustomStatusProgressbar is now a Singleton
